Change absolute paths to relative for index
Currently the index tar that is made is tarred with the absolute
filepaths of all the .cabal files that go into the tar. While this works
find on Unix-y machines, for some reason this introduces problems on
Windows where the absolute path turns into "C: \Users\..." (with an
extra space there), which makes yackage unusable on Windows. By changing
the paths of the .cabal files to be relative, we nicely sidestep this
problem.